fix: change in the type of the return of va_arg

+++ function_numerics.c
@@ -8,7 +8,7 @@
  */
 int op_integer(va_list list)
 {
-	int num, temp;
+	long int num, temp;
 	int nbytes, count = 0;
 	char *str;
 
@@ -51,7 +51,7 @@ int op_integer(va_list list)
  */
 int op_binary(va_list list)
 {
-	int num, temp;
+	long int num, temp;
 	int nbytes, i = 0, count = 0;
 	char *str;
